requirements.txt
# Base
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0

# Linting
flake8==3.8.3
flake8-django==1.1.1
isort==5.1.4

# Testing
django-nose==1.4.6
coverage==5.2
